e2e(ssr): extracted `seed-data` and `storage`
